# target/classes/swagger/produto-odonto.yaml
swagger: '2.0'
info:
  description: Serviço de Produto do Odonto
  version: 1.0.0
  title: produto-odonto
  x-ibm-name: produto-odonto
host: api.us.apiconnect.ibmcloud.com
basePath: /v1/produto-odonto
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  '/carteirinha/cliente/{cpf}':
    get:
      summary: Pesquisar carteirinha de um cliente
      tags:
        - produto
      operationId: '/carteirinha/cliente/{cpf}'
      parameters:
        - name: cpf
          in: path
          description: CPF do Segurado
          required: true
          type: integer
          format: int64
        - name: Token-Unimed
          type: string
          required: true
          in: header
          description: JWT
      responses:
        '200':
          description: Operação com sucesso
          schema:
            $ref: '#/definitions/pesquisarCarteirinhaType'
        '400':
          description: Informações inválidas
  '/carteirinha/{numero}/cpf/{cpf}':
    get:
      summary: Consultar detalhes de uma carteirinha
      tags:
        - produto
      operationId: '/carteirinha/{numero}/cpf/{cpf}'
      parameters:
        - name: numero
          in: path
          description: Número da Carteirinha
          required: true
          type: integer
          format: int64
        - name: cpf
          in: path
          description: CPF do Segurado
          required: true
          type: integer
          format: int64
        - name: Token-Unimed
          type: string
          required: true
          in: header
          description: JWT
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/consultarCarteirinhaType'
        '400':
          description: Informações inválidas
definitions:
  pesquisarCarteirinhaType:
    allOf:
      - $ref: '#/definitions/status_statusProcessamentoType'
      - type: object
        properties:
          resultado:
            $ref: '#/definitions/pesquisarCarteirinhaResultado'
  status_statusProcessamentoType:
    type: object
    properties:
      codigo:
        type: integer
        format: int32
      descricao:
        type: string
      detalhe:
        type: string
      mensagens:
        $ref: '#/definitions/status_mensagensType'
    required:
      - codigo
      - descricao
  status_mensagensType:
    type: object
    properties:
      mensagem:
        type: array
        items:
          $ref: '#/definitions/status_mensagem'
  status_mensagemType:
    type: object
    properties:
      descricao:
        type: string
      detalhe:
        type: string
  pesquisarCarteirinhaResultado:
    type: object
    properties:
      listaCarteirinha:
        $ref: '#/definitions/listaDadosPesquisarCarteirinha'
  status_mensagem:
    allOf:
      - $ref: '#/definitions/status_mensagemType'
      - type: object
  listaDadosPesquisarCarteirinha:
    type: object
    properties:
      carteirinha:
        type: array
        items:
          $ref: '#/definitions/dadosPesquisarCarteirinha'
  dadosPesquisarCarteirinha:
    type: object
    properties:
      produto:
        type: string
      numero:
        type: integer
        format: int32
      inicioVigencia:
        type: string
        format: date
      beneficiario:
        $ref: '#/definitions/dadosBeneficiario'
      status:
        type: string
  consultarCarteirinhaType:
    allOf:
      - $ref: '#/definitions/status_statusProcessamentoType'
      - type: object
        properties:
          resultado:
            $ref: '#/definitions/consultarCarteirinhaResultado'
  consultarCarteirinhaResultado:
    type: object
    properties:
      beneficiario:
        $ref: '#/definitions/dadosBeneficiario'
  dadosBeneficiario:
    type: object
    properties:
      nome:
        type: string
      numero:
        type: integer
        format: int64
      dataNascimento:
        type: string
        format: date
      situacao:
        type: string
      numeroCns:
        type: integer
        format: int64
      dataInclusaoPlano:
        type: string
        format: date
      tipoContrato:
        type: string
      categoriaTipoContrato:
        type: string
        enum:
          - Adesao
          - Empresarial
          - individual
      numeroContrato:
        type: integer
        format: int64
      nomeEmpresa:
        type: string
      endereco:
        $ref: '#/definitions/endereco'
      listaPlano:
        $ref: '#/definitions/listaDadosPlano'
      listaCobertura:
        $ref: '#/definitions/listaCobertura'
      listaBeneficio:
        $ref: '#/definitions/listaBeneficio'
      infoComplementar:
        $ref: '#/definitions/dadosInfoComplementar'
      listaDependente:
        $ref: '#/definitions/listaDadosDependente'
      tipo:
        $ref: '#/definitions/enumTipoBeneficiarioType'
  enumTipoBeneficiarioType:
    type: string
    enum:
      - T
      - D
  listaDadosDependente:
    type: object
    properties:
      dependente:
        type: array
        items:
          $ref: '#/definitions/dadosDependente'
  dadosDependente:
    type: object
    properties:
      nome:
        type: string
      dataNascimento:
        type: string
        format: date
      listaPlano:
        $ref: '#/definitions/listaDadosPlano'
      listaCobertura:
        $ref: '#/definitions/listaCobertura'
      listaBeneficio:
        $ref: '#/definitions/listaBeneficio'
      infoComplementar:
        $ref: '#/definitions/dadosInfoComplementar'
      numero:
        type: string
      numeroCNS:
        type: string
      situacao:
        type: string
  listaDadosPlano:
    type: object
    properties:
      plano:
        type: array
        items:
          $ref: '#/definitions/dadosPlano'
  dadosPlano:
    type: object
    properties:
      codigo:
        type: string
      nome:
        type: string
      codigoRegistroAns:
        type: integer
        format: int64
      abrangencia:
        type: string
      nomeProduto:
        type: string
      descricaoRegulamentacao:
        type: string
      tipoAcomodacao:
        type: string
      segmentacaoAssistencial:
        type: string
      tipoContratacao:
        type: string
      dataInicio:
        type: string
        format: date
      validadeCarteirinha:
        type: string
        format: date
      viaCarteirinha:
        type: string
      dataCpt:
        type: string
        format: date
      dataCarenciaUrgenciaEmergencia:
        type: string
        format: date
      dataCarenciaConsultaExameBasico:
        type: string
        format: date
      dataCarenciaTerapiaExamesEspeciais:
        type: string
        format: date
      carencias:
        $ref: '#/definitions/carenciasType'
  carenciaType:
    type: object
    properties:
      especialidade:
        $ref: '#/definitions/especialidadeType'
      dataFim:
        type: string
        format: date
  especialidadeType:
    type: object
    properties:
      nome:
        type: string
  carenciasType:
    type: array
    items:
      $ref: '#/definitions/carenciaType'
  listaCobertura:
    type: object
    properties:
      cobertura:
        type: array
        items:
          $ref: '#/definitions/dadosCobertura'
  dadosCobertura:
    type: object
    properties:
      codigo:
        type: string
      nome:
        type: string
      quantidadeLimiteUso:
        type: integer
        format: int32
      flagReembolso:
        type: string
  listaBeneficio:
    type: object
    properties:
      beneficio:
        type: array
        items:
          $ref: '#/definitions/dadosBeneficio'
  dadosBeneficio:
    type: object
    properties:
      codigo:
        type: string
      nome:
        type: string
      dataInicioVigencia:
        type: string
        format: date
      dataFimVigencia:
        type: string
        format: date
  dadosInfoComplementar:
    type: object
    properties:
      codigoRegistroOperadora:
        type: string
      nomeOperadora:
        type: string
      numeroSac:
        type: integer
        format: int64
      numeroAns:
        type: integer
        format: int64
      siteAns:
        type: string
  endereco:
    type: object
    properties:
      cep:
        type: string
      tipoLogradouro:
        type: string
      logradouro:
        type: string
      numero:
        type: integer
        format: int64
      complemento:
        type: string
      bairro:
        type: string
      cidade:
        type: string
      estado:
        $ref: '#/definitions/enumUf'
  enumUf:
    example: AC
    type: string
    enum:
      - AC AL AM AP BA CE DF ES GO MA MG MS MT PA PB PE PI PR RJ RN RO RR RS SC SE SP TO
...
security:
  - clientIdHeader: []
securityDefinitions:
  clientIdHeader:
    type: apiKey
    description: ''
    in: header
    name: X-IBM-Client-Id